+++ contracts/Vader.sol
@@ -8,27 +8,28 @@ import "./interfaces/iUSDV.sol";
 import "./interfaces/iROUTER.sol";
 
 contract Vader is iERC20 {
-
     // ERC-20 Parameters
-    string public override name; string public override symbol;
-    uint public override decimals; uint public override totalSupply;
+    string public override name;
+    string public override symbol;
+    uint256 public override decimals;
+    uint256 public override totalSupply;
 
     // ERC-20 Mappings
-    mapping(address => uint) private _balances;
-    mapping(address => mapping(address => uint)) private _allowances;
+    mapping(address => uint256) private _balances;
+    mapping(address => mapping(address => uint256)) private _allowances;
 
     // Parameters
     bool private inited;
     bool public emitting;
     bool public minting;
-    uint _1m;
-    uint public baseline;
-    uint public emissionCurve;
-    uint public maxSupply;
-    uint public secondsPerEra;
-    uint public currentEra;
-    uint public nextEraTime;
-    uint public feeOnTransfer;
+    uint256 _1m;
+    uint256 public baseline;
+    uint256 public emissionCurve;
+    uint256 public maxSupply;
+    uint256 public secondsPerEra;
+    uint256 public currentEra;
+    uint256 public nextEraTime;
+    uint256 public feeOnTransfer;
 
     address public VETHER;
     address public USDV;
@@ -37,7 +38,7 @@ contract Vader is iERC20 {
     address public rewardAddress;
     address public DAO;
 
-    event NewEra(uint currentEra, uint nextEraTime, uint emission);
+    event NewEra(uint256 currentEra, uint256 nextEraTime, uint256 emission);
 
     // Only DAO can execute
     modifier onlyDAO() {
@@ -49,17 +50,18 @@ contract Vader is iERC20 {
         require(isMature(), "No flash");
         _;
     }
-    function isMature() public view returns(bool){
+
+    function isMature() public view returns (bool) {
         return iUSDV(USDV).isMature();
     }
 
     //=====================================CREATION=========================================//
     // Constructor
     constructor() {
-        name = 'VADER PROTOCOL TOKEN';
-        symbol = 'VADER';
+        name = "VADER PROTOCOL TOKEN";
+        symbol = "VADER";
         decimals = 18;
-        _1m = 10**6 * 10 ** decimals; //1m
+        _1m = 10**6 * 10**decimals; //1m
         baseline = _1m;
         totalSupply = 0;
         maxSupply = 2 * _1m;
@@ -70,8 +72,13 @@ contract Vader is iERC20 {
         DAO = msg.sender;
         burnAddress = 0x0111011001100001011011000111010101100101;
     }
+
     // Can only be called once
-    function init(address _vether, address _USDV, address _utils) external {
+    function init(
+        address _vether,
+        address _USDV,
+        address _utils
+    ) external {
         require(inited == false);
         inited = true;
         VETHER = _vether;
@@ -81,50 +88,67 @@ contract Vader is iERC20 {
     }
 
     //========================================iERC20=========================================//
-    function balanceOf(address account) external view override returns (uint) {
+    function balanceOf(address account) external view override returns (uint256) {
         return _balances[account];
     }
-    function allowance(address owner, address spender) public view virtual override returns (uint) {
+
+    function allowance(address owner, address spender) public view virtual override returns (uint256) {
         return _allowances[owner][spender];
     }
+
     // iERC20 Transfer function
-    function transfer(address recipient, uint amount) external virtual override returns (bool) {
+    function transfer(address recipient, uint256 amount) external virtual override returns (bool) {
         _transfer(msg.sender, recipient, amount);
         return true;
     }
+
     // iERC20 Approve, change allowance functions
-    function approve(address spender, uint amount) external virtual override returns (bool) {
+    function approve(address spender, uint256 amount) external virtual override returns (bool) {
         _approve(msg.sender, spender, amount);
         return true;
     }
-    function _approve(address owner, address spender, uint amount) internal virtual {
+
+    function _approve(
+        address owner,
+        address spender,
+        uint256 amount
+    ) internal virtual {
         require(owner != address(0), "sender");
         require(spender != address(0), "spender");
         _allowances[owner][spender] = amount;
         emit Approval(owner, spender, amount);
     }
-    
+
     // iERC20 TransferFrom function
-    function transferFrom(address sender, address recipient, uint amount) external virtual override returns (bool) {
+    function transferFrom(
+        address sender,
+        address recipient,
+        uint256 amount
+    ) external virtual override returns (bool) {
         _transfer(sender, recipient, amount);
         _approve(sender, msg.sender, _allowances[sender][msg.sender] - amount);
         return true;
     }
 
     // TransferTo function
-    // Risks: User can be phished, or tx.origin may be deprecated, optionality should exist in the system. 
-    function transferTo(address recipient, uint amount) external virtual override returns (bool) {
+    // Risks: User can be phished, or tx.origin may be deprecated, optionality should exist in the system.
+    function transferTo(address recipient, uint256 amount) external virtual override returns (bool) {
         _transfer(tx.origin, recipient, amount);
         return true;
     }
 
     // Internal transfer function
-    function _transfer(address sender, address recipient, uint amount) internal virtual {
+    function _transfer(
+        address sender,
+        address recipient,
+        uint256 amount
+    ) internal virtual {
         require(sender != address(0), "sender");
         require(recipient != address(this), "recipient");
         _balances[sender] -= amount;
-        uint _fee = iUTILS(UTILS).calcPart(feeOnTransfer, amount);  // Critical functionality
-        if(_fee >= 0 && _fee <= amount){                            // Stops reverts if UTILS corrupted
+        uint256 _fee = iUTILS(UTILS).calcPart(feeOnTransfer, amount); // Critical functionality
+        if (_fee >= 0 && _fee <= amount) {
+            // Stops reverts if UTILS corrupted
             amount -= _fee;
             _burn(msg.sender, _fee);
         }
@@ -132,26 +156,30 @@ contract Vader is iERC20 {
         emit Transfer(sender, recipient, amount);
         _checkEmission();
     }
+
     // Internal mint (upgrading and daily emissions)
-    function _mint(address account, uint amount) internal virtual {
+    function _mint(address account, uint256 amount) internal virtual {
         require(account != address(0), "recipient");
-        if((totalSupply + amount) >= maxSupply){
-            amount = maxSupply - totalSupply;       // Safety, can't mint above maxSupply
+        if ((totalSupply + amount) >= maxSupply) {
+            amount = maxSupply - totalSupply; // Safety, can't mint above maxSupply
         }
         totalSupply += amount;
         _balances[account] += amount;
         emit Transfer(address(0), account, amount);
     }
+
     // Burn supply
-    function burn(uint amount) public virtual override {
+    function burn(uint256 amount) public virtual override {
         _burn(msg.sender, amount);
     }
-    function burnFrom(address account, uint amount) external virtual override {
-        uint decreasedAllowance = allowance(account, msg.sender) - amount;
+
+    function burnFrom(address account, uint256 amount) external virtual override {
+        uint256 decreasedAllowance = allowance(account, msg.sender) - amount;
         _approve(account, msg.sender, decreasedAllowance);
         _burn(account, amount);
     }
-    function _burn(address account, uint amount) internal virtual {
+
+    function _burn(address account, uint256 amount) internal virtual {
         require(account != address(0), "address err");
         _balances[account] -= amount;
         totalSupply -= amount;
@@ -161,82 +189,95 @@ contract Vader is iERC20 {
     //=========================================DAO=========================================//
     // Can start
     function flipEmissions() external onlyDAO {
-        if(emitting){
+        if (emitting) {
             emitting = false;
         } else {
             emitting = true;
         }
     }
+
     // Can stop
     function flipMinting() external onlyDAO {
-        if(minting){
+        if (minting) {
             minting = false;
         } else {
             minting = true;
         }
     }
+
     // Can set params
-    function setParams(uint newEra, uint newCurve) external onlyDAO {
+    function setParams(uint256 newEra, uint256 newCurve) external onlyDAO {
         secondsPerEra = newEra;
         emissionCurve = newCurve;
     }
+
     // Can set reward address
     function setRewardAddress(address newAddress) external onlyDAO {
         rewardAddress = newAddress;
     }
+
     // Can change UTILS
     function changeUTILS(address newUTILS) external onlyDAO {
         require(newUTILS != address(0), "address err");
         UTILS = newUTILS;
     }
+
     // Can change DAO
     function changeDAO(address newDAO) external onlyDAO {
         require(newDAO != address(0), "address err");
         DAO = newDAO;
     }
+
     // Can purge DAO
-    function purgeDAO() external onlyDAO{
+    function purgeDAO() external onlyDAO {
         DAO = address(0);
     }
 
-   //======================================EMISSION========================================//
+    //======================================EMISSION========================================//
     // Internal - Update emission function
     function _checkEmission() private {
-        if ((block.timestamp >= nextEraTime) && emitting) {                                // If new Era and allowed to emit
-            currentEra += 1;                                                               // Increment Era
-            nextEraTime = block.timestamp + secondsPerEra;                                 // Set next Era time
-            uint _emission = getDailyEmission();                                           // Get Daily Dmission
-            _mint(rewardAddress, _emission);                                               // Mint to the Rewad Address
-            feeOnTransfer = iUTILS(UTILS).getFeeOnTransfer(totalSupply, maxSupply);        // UpdateFeeOnTransfer
-            if(feeOnTransfer > 1000){feeOnTransfer = 1000;}                                // Max 10% if UTILS corrupted
-            emit NewEra(currentEra, nextEraTime, _emission);                               // Emit Event
+        if ((block.timestamp >= nextEraTime) && emitting) {
+            // If new Era and allowed to emit
+            currentEra += 1; // Increment Era
+            nextEraTime = block.timestamp + secondsPerEra; // Set next Era time
+            uint256 _emission = getDailyEmission(); // Get Daily Dmission
+            _mint(rewardAddress, _emission); // Mint to the Rewad Address
+            feeOnTransfer = iUTILS(UTILS).getFeeOnTransfer(totalSupply, maxSupply); // UpdateFeeOnTransfer
+            if (feeOnTransfer > 1000) {
+                feeOnTransfer = 1000;
+            } // Max 10% if UTILS corrupted
+            emit NewEra(currentEra, nextEraTime, _emission); // Emit Event
         }
     }
+
     // Calculate Daily Emission
-    function getDailyEmission() public view returns (uint) {
-        uint _adjustedMax;
-        if(totalSupply <= baseline){ // If less than 1m, then adjust cap down
+    function getDailyEmission() public view returns (uint256) {
+        uint256 _adjustedMax;
+        if (totalSupply <= baseline) {
+            // If less than 1m, then adjust cap down
             _adjustedMax = (maxSupply * totalSupply) / baseline; // 2m * 0.5m / 1m = 2m * 50% = 1.5m
         } else {
-            _adjustedMax = maxSupply;  // 2m
+            _adjustedMax = maxSupply; // 2m
         }
-        return (_adjustedMax - totalSupply) / (emissionCurve); // outstanding / curve 
+        return (_adjustedMax - totalSupply) / (emissionCurve); // outstanding / curve
     }
 
     //======================================ASSET MINTING========================================//
     // VETHER Owners to Upgrade
-    function upgrade(uint amount) external {
+    function upgrade(uint256 amount) external {
         require(iERC20(VETHER).transferFrom(msg.sender, burnAddress, amount));
         _mint(msg.sender, amount);
     }
+
     // Directly redeem back to VADER (must have sent USDV first)
-    function redeem() external returns (uint redeemAmount){
+    function redeem() external returns (uint256 redeemAmount) {
         return redeemToMember(msg.sender);
     }
+
     // Redeem on behalf of member (must have sent USDV first)
-    function redeemToMember(address member) public flashProof returns (uint redeemAmount){
-        if(minting){
-            uint _amount = iERC20(USDV).balanceOf(address(this)); 
+    function redeemToMember(address member) public flashProof returns (uint256 redeemAmount) {
+        if (minting) {
+            uint256 _amount = iERC20(USDV).balanceOf(address(this));
             iERC20(USDV).burn(_amount);
             redeemAmount = iROUTER(iUSDV(USDV).ROUTER()).getVADERAmount(_amount); // Critical pricing functionality
             _mint(member, redeemAmount);
