+++ contracts/interfaces/iSYNTH.sol
@@ -2,6 +2,7 @@
 pragma solidity 0.8.3;
 
 interface iSYNTH {
-    function mint(address account, uint amount) external;
-    function TOKEN() external view returns(address);
+    function mint(address account, uint256 amount) external;
+
+    function TOKEN() external view returns (address);
 }
